+++ ui/src/side_nav/containers/SideNav.js
@@ -13,6 +13,8 @@ import {DEFAULT_HOME_PAGE} from 'shared/constants'
 
 const {bool, shape, string} = PropTypes
 
+const V_NUMBER = VERSION // eslint-disable-line no-undef
+
 const SideNav = React.createClass({
   propTypes: {
     params: shape({
@@ -40,12 +42,14 @@ const SideNav = React.createClass({
     return isHidden
       ? null
       : <NavBar location={location}>
-          <Link
-            to={`${sourcePrefix}/${DEFAULT_HOME_PAGE}`}
-            className="sidebar__logo"
-          >
-            <span className="icon cubo-uniform" />
-          </Link>
+          <div className="sidebar--item">
+            <Link
+              to={`${sourcePrefix}/${DEFAULT_HOME_PAGE}`}
+              className="sidebar--square sidebar--logo"
+            >
+              <span className="sidebar--icon icon cubo-uniform" />
+            </Link>
+          </div>
           <NavBlock icon="cubo-node" link={`${sourcePrefix}/hosts`}>
             <NavHeader link={`${sourcePrefix}/hosts`} title="Host List" />
           </NavBlock>
@@ -80,11 +84,27 @@ const SideNav = React.createClass({
               title="Configuration"
             />
           </NavBlock>
-          {showLogout
-            ? <NavBlock icon="user" className="sidebar__square-last">
-                <NavHeader useAnchor={true} link={logoutLink} title="Logout" />
-              </NavBlock>
-            : null}
+          <div className="sidebar--bottom">
+            <div className="sidebar--item">
+              <div className="sidebar--square">
+                <span className="sidebar--icon icon zap" />
+              </div>
+              <div className="sidebar-menu">
+                <div className="sidebar-menu--heading">
+                  Version: {V_NUMBER}
+                </div>
+              </div>
+            </div>
+            {showLogout
+              ? <NavBlock icon="user" className="sidebar--item-last">
+                  <NavHeader
+                    useAnchor={true}
+                    link={logoutLink}
+                    title="Logout"
+                  />
+                </NavBlock>
+              : null}
+          </div>
         </NavBar>
   },
 })