+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v1.3.4.0 [unreleased]
 ### Bug Fixes
 1. [#1612](https://github.com/influxdata/chronograf/pull/1612): Prevent users from being able to write to internal system databases
+1. [#1655](https://github.com/influxdata/chronograf/pull/1655): Add more than one color to Line+Stat graphs
 
 ### Features
 1. [#1645](https://github.com/influxdata/chronograf/pull/1645): Add Auth0 as a supported OAuth2 provider
@@ -8,6 +9,17 @@
 ### UI Improvements
 1. [#1644](https://github.com/influxdata/chronograf/pull/1644): Redesign Alerts History table to have sticky headers
 1. [#1581](https://github.com/influxdata/chronograf/pull/1581): Refresh template variable values on dashboard page load
+1. [#1612](https://github.com/influxdata/chronograf/pull/1612): Prevent users from being able to write to internal system databases
+1. [#1655](https://github.com/influxdata/chronograf/pull/1655): Add version number item to the navbar
+1. [#1655](https://github.com/influxdata/chronograf/pull/1655): Redesign dashboards table and sort alphabetically by name
+1. [#1655](https://github.com/influxdata/chronograf/pull/1655): Redesign navbar to be consistent with navbar in Branding Documentation
+
+## v1.3.3.3 [2017-06-21]
+### Bug Fixes
+1. [1651](https://github.com/influxdata/chronograf/pull/1651): Add back in x and y axes and revert some style changes on Line + Single Stat graphs
+
+## v1.3.3.2 [2017-06-21]
+### Bug Fixes
 
 ## v1.3.3.3 [2017-06-21]
 ### Bug Fixes